Prepare message population.

+++ Lettera/app/src/main/java/org/purple/lettera/FolderElement.java
@@ -35,5 +35,5 @@ public class FolderElement
     public int m_is_regular_folder = 2;
     public int m_message_count = 0;
     public int m_new_message_count = 0;
-    public int m_oid = -1;
+    public long m_oid = -1;
 }
